+++ src/app/app.module.ts
@@ -16,11 +16,11 @@ import { SharedModule } from './shared/shared.module';
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule,
     ProductsModule,
     CartModule,
     BrowserAnimationsModule,
-    SharedModule
+    SharedModule,
+    AppRoutingModule,
   ],
   bootstrap: [AppComponent]
 })