+++ src/app/cart/models/cart-model.ts
@@ -2,4 +2,4 @@ import { ProductModel } from "src/app/products/models/product-model";
 
 export interface CartModel extends ProductModel {
     quantity: number;
-} 
+}