+++ src/app/products/products.module.ts
@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { ProductComponent } from './components/product/product.component';
-import { ProductService } from './services/product.service';
 import { SharedModule } from '../shared/shared.module';
 import { RouterModule } from '@angular/router';
 
@@ -15,7 +13,6 @@ import { RouterModule } from '@angular/router';
     ProductListComponent,
   ],
   imports: [
-    CommonModule,
     SharedModule,
     RouterModule
   ],